pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = "Discord bot powered by Google's Agent Development Kit (ADK)"
authors = ["Younes <younes.belataris@appchoose.io>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
"discord.py" = "^2.5.2"
google-adk = "*"
python-dotenv = "*"

[tool.poetry.scripts]
discord-bot = "src.bot:main"
run-bot = "src.bot:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
